python/requirements.txt
requests
python-dotenv
chromadb
openai
fastapi
uvicorn 
# Python dependencies for TradeNexus RAG Backend

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# AI/ML dependencies
openai==1.3.8
chromadb==0.4.18

# HTTP client
requests==2.31.0

# CORS support
fastapi-cors==0.0.6

# Optional: For better performance
python-multipart==0.0.6